+++ controllers/ban.go
@@ -1,0 +1,62 @@
+package controllers
+
+import (
+	"net/http"
+	"strings"
+
+	lru "github.com/hashicorp/golang-lru/v2"
+	"github.com/labstack/echo/v4"
+	"github.com/rs/zerolog"
+	"gitlab.com/etke.cc/buscarron/metrics"
+)
+
+type banner struct {
+	store *lru.Cache[string, struct{}]
+	log   *zerolog.Logger
+}
+
+var donotban = []string{"/favicon.ico", "/robots.txt", "/metrics", "/_domain", "/_health"}
+
+// NewBanner creates banner
+func NewBanner(size int, banlist []string, log *zerolog.Logger) *banner {
+	store, err := lru.New[string, struct{}](size)
+	if err != nil {
+		log.Error().Err(err).Msg("cannot init cache")
+	}
+	log.Info().Strs("ids", banlist).Msg("preemptive banning...")
+	for _, ip := range banlist {
+		ip = strings.TrimSpace(ip)
+		store.Add(ip, struct{}{})
+		metrics.BanUser("persistent", "-")
+	}
+
+	return &banner{store, log}
+}
+
+// Ban by request data
+func (b *banner) Ban(c echo.Context, reason string) {
+	for _, exclude := range donotban {
+		if c.Request().URL.Path == exclude {
+			return
+		}
+	}
+	ip := c.RealIP()
+	b.store.Add(ip, struct{}{})
+	metrics.BanUser(reason, c.Param("name"))
+
+	b.log.Info().Str("method", c.Request().Method).Str("path", c.Request().URL.Path).Str("ip", ip).Msg("has been banned")
+}
+
+func (b *banner) Middleware() echo.MiddlewareFunc {
+	return func(next echo.HandlerFunc) echo.HandlerFunc {
+		return func(c echo.Context) error {
+			ip := c.RealIP()
+			if b.store.Contains(c.RealIP()) {
+				metrics.BanRequest()
+				b.log.Debug().Str("method", c.Request().Method).Str("path", c.Request().URL.Path).Str("ip", ip).Msg("banned request attempt")
+				return c.NoContent(http.StatusTooManyRequests)
+			}
+			return next(c)
+		}
+	}
+}